refactor: replace directory with files
