--- src/apis/constants/member/index.ts
@@ -1,23 +1,0 @@
-const MEMBER_API_COMMON_SEGMENT = '/member';
-
-const MEMBER_API_SEGMENTS = {
-  LOGIN: '/login',
-  SCHOOL_EMAIL: '/code',
-  SIGNUP: '/signup',
-  REMAKE_TOKEN: '/remake/token',
-  CHECK_NICKNAME: '/check-nickname',
-  LOGOUT: '/logout',
-  DELETE: '/delete',
-};
-
-const MEMBER_API_PATHS = {
-  LOGIN: '/member/login',
-  SCHOOL_EMAIL: '/member/school-email/code',
-  SIGNUP: '/member/signup',
-  REMAKE_TOKEN: '/member/remake/token',
-  CHECK_NICKNAME: '/member/check-nickname',
-  LOGOUT: '/member/logout',
-  DELETE: '/member/delete',
-};
-
-export default MEMBER_API_PATHS;
