+++ src/apis/member.ts
@@ -1,5 +1,28 @@
-import { HTTP_METHODS } from '../constants';
-import MEMBER_API_PATHS from '../constants/member';
+import { HTTP_METHODS } from './constants';
+
+const MEMBER_API_COMMON_SEGMENT = '/member';
+
+const MEMBER_API_SEGMENTS = {
+  LOGIN: '/login',
+  SCHOOL_EMAIL: '/code',
+  SIGNUP: '/signup',
+  REMAKE_TOKEN: '/remake/token',
+  CHECK_NICKNAME: '/check-nickname',
+  LOGOUT: '/logout',
+  DELETE: '/delete',
+};
+
+const MEMBER_API_PATHS = {
+  LOGIN: '/member/login',
+  SCHOOL_EMAIL: '/member/school-email/code',
+  SIGNUP: '/member/signup',
+  REMAKE_TOKEN: '/member/remake/token',
+  CHECK_NICKNAME: '/member/check-nickname',
+  LOGOUT: '/member/logout',
+  DELETE: '/member/delete',
+};
+
+export default MEMBER_API_PATHS;
 
 export const getAuth = async (idToken: string) => {
   try {
