+++ src/pages/Callback/useTokenExchange/index.tsx
@@ -1,8 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { HTTP_HEADERS, HTTP_METHODS } from '../../../apis/constants';
-import getIdToken from '../../../apis/auth';
-import { getAuth } from '../../../apis/member';
 import useAuthStore from '../../../store/authStore';
 
 interface AuthType {
